Reuse Stable price impact concern for MetaStable and StablePhantom pools
