+++ balancer-js/src/modules/pools/pool-types/concerns/metaStable/priceImpact.spec.ts
@@ -1,10 +1,10 @@
 // yarn test:only src/modules/pools/pool-types/concerns/metaStable/priceImpact.spec.ts
 import { expect } from 'chai';
-import { MetaStablePoolPriceImpact } from '@/modules/pools/pool-types/concerns/metaStable/priceImpact.concern';
+import { StablePoolPriceImpact } from '@/modules/pools/pool-types/concerns/stable/priceImpact.concern';
 import pools_14717479 from '@/test/lib/pools_14717479.json';
 import { Pool } from '@/types';
 
-const priceImpactCalc = new MetaStablePoolPriceImpact();
+const priceImpactCalc = new StablePoolPriceImpact();
 const wstETHwETH =
   '0x32296969ef14eb0c6d29669c550d4a0449130230000200000000000000000080';
 
@@ -12,19 +12,16 @@ const pool = pools_14717479.find(
   (pool) => pool.id == wstETHwETH
 ) as unknown as Pool;
 
-const tokenAmounts = [
-  BigInt('629870162919981039400158'),
-  BigInt('615159929697'),
-];
+const tokenAmounts = ['629870162919981039400158', '615159929697'];
 
 describe('metastable pool price impact', () => {
   context('bpt zero price impact', () => {
     it('non-proportional case', () => {
       const bptZeroPriceImpact = priceImpactCalc.bptZeroPriceImpact(
         pool,
-        tokenAmounts
+        tokenAmounts.map(BigInt)
       );
-      expect(bptZeroPriceImpact.toString()).to.eq('662816325116386208862285');
+      expect(bptZeroPriceImpact.toString()).to.eq('662816325116386209174659');
     });
     it('proportional case', () => {
       // This tokenAmounts vector is proportional to the balances
@@ -34,12 +31,11 @@ describe('metastable pool price impact', () => {
         BigInt('813913487516879908953'),
         BigInt('854410030026808373669'),
       ];
-
       const bptZeroPriceImpact = priceImpactCalc.bptZeroPriceImpact(
         pool,
         proportionalTokenAmounts
       );
-      expect(bptZeroPriceImpact.toString()).to.eq('1696871032806568300470');
+      expect(bptZeroPriceImpact.toString()).to.eq('1696871032806568300873');
     });
   });
 
@@ -47,11 +43,11 @@ describe('metastable pool price impact', () => {
     it('calculate price impact', () => {
       const priceImpact = priceImpactCalc.calcPriceImpact(
         pool,
-        tokenAmounts.map((amount) => amount.toString()),
+        tokenAmounts,
         '660816325116386208862285',
         true
       );
-      expect(priceImpact.toString()).to.eq('3017427187914862');
+      expect(priceImpact.toString()).to.eq('3017427187914863');
     });
   });
 });
