+++ balancer-js/src/modules/pools/pool-types/concerns/stable/priceImpact.concern.ts
@@ -6,6 +6,7 @@ import { Pool } from '@/types';
 import { bptSpotPrice } from '@/lib/utils/stableMathHelpers';
 import { parsePoolInfo } from '@/lib/utils';
 
+// Note: this concern is used by Stable, MetaStable and StablePhantom pools
 export class StablePoolPriceImpact implements PriceImpactConcern {
   /**
    * Calculates the BPT return amount when investing with no price impact.
@@ -14,26 +15,31 @@ export class StablePoolPriceImpact implements PriceImpactConcern {
    * @returns { bigint } BPT amount.
    */
   bptZeroPriceImpact(pool: Pool, tokenAmounts: bigint[]): bigint {
-    if (tokenAmounts.length !== pool.tokensList.length)
-      throw new BalancerError(BalancerErrorCode.INPUT_LENGTH_MISMATCH);
-
     // upscales amp, swapfee, totalshares
     const {
       ampWithPrecision,
+      scalingFactorsWithoutBpt,
       totalSharesEvm,
-      scalingFactorsRaw,
-      upScaledBalances,
+      upScaledBalancesWithoutBpt,
     } = parsePoolInfo(pool);
 
+    // Check against array without BPT because concern is being reused by stablePhantom pools
+    // Stable and MetaStable pools don't have BPT in the token list, so array without BPT is the same as the original array
+    if (tokenAmounts.length !== upScaledBalancesWithoutBpt.length)
+      throw new BalancerError(BalancerErrorCode.INPUT_LENGTH_MISMATCH);
+
     let bptZeroPriceImpact = BZERO;
-    for (let i = 0; i < upScaledBalances.length; i++) {
+    for (let i = 0; i < upScaledBalancesWithoutBpt.length; i++) {
       const price = bptSpotPrice(
         ampWithPrecision,
-        upScaledBalances,
+        upScaledBalancesWithoutBpt,
         totalSharesEvm,
         i
       );
-      const amountUpscaled = _upscale(tokenAmounts[i], scalingFactorsRaw[i]);
+      const amountUpscaled = _upscale(
+        tokenAmounts[i],
+        scalingFactorsWithoutBpt[i]
+      );
       const newTerm = (price * amountUpscaled) / ONE;
       bptZeroPriceImpact += newTerm;
     }
