+++ balancer-js/src/modules/pools/pool-types/concerns/stablePhantom/priceImpact.spec.ts
@@ -1,11 +1,10 @@
 // yarn test:only src/modules/pools/pool-types/concerns/stablePhantom/priceImpact.spec.ts
 import { expect } from 'chai';
-import { StablePhantomPriceImpact } from '@/modules/pools/pool-types/concerns/stablePhantom/priceImpact.concern';
+import { StablePoolPriceImpact } from '@/modules/pools/pool-types/concerns/stable/priceImpact.concern';
 import pools_14717479 from '@/test/lib/pools_14717479.json';
 import { Pool } from '@/types';
-import { parseFixed } from '@/lib/utils';
 
-const priceImpactCalc = new StablePhantomPriceImpact();
+const priceImpactCalc = new StablePoolPriceImpact();
 const bbaUSDPoolId =
   '0x7b50775383d3d6f0215a8f290f2c9e2eebbeceb20000000000000000000000fe';
 
@@ -14,9 +13,9 @@ const pool = pools_14717479.find(
 ) as unknown as Pool;
 
 const tokenAmounts = [
-  BigInt('6298701629199810399876'),
-  BigInt('615159929697'),
-  BigInt('101515992969778'),
+  '6298701629199810399876',
+  '615159929697',
+  '101515992969778',
 ];
 
 describe('phantomStable pool price impact', () => {
@@ -24,14 +23,14 @@ describe('phantomStable pool price impact', () => {
     it('non-proportional case', () => {
       const bptZeroPriceImpact = priceImpactCalc.bptZeroPriceImpact(
         pool,
-        tokenAmounts
+        tokenAmounts.map(BigInt)
       );
-      expect(bptZeroPriceImpact.toString()).to.eq('6310741387055771004078');
+      expect(bptZeroPriceImpact.toString()).to.eq('6294084206629046579738');
     });
 
     it('proportional case', () => {
       // the correct return value is totalShares times 0.01
-      const tokenAmounts = [
+      const proportionalTokenAmounts = [
         BigInt('831191821406963569140405'),
         BigInt('851842896587052519012488'),
         BigInt('906277003015102397681882'),
@@ -39,12 +38,10 @@ describe('phantomStable pool price impact', () => {
 
       const bptZeroPriceImpact = priceImpactCalc.bptZeroPriceImpact(
         pool,
-        tokenAmounts
+        proportionalTokenAmounts
       );
-      console.log(bptZeroPriceImpact.toString());
-      console.log(parseFixed(pool.totalShares, 18).toString());
 
-      expect(bptZeroPriceImpact.toString()).to.eq('2584652218704385059205928');
+      expect(bptZeroPriceImpact.toString()).to.eq('2584652218704385060046703');
     });
   });
 
@@ -52,11 +49,11 @@ describe('phantomStable pool price impact', () => {
     it('calculate price impact', () => {
       const priceImpact = priceImpactCalc.calcPriceImpact(
         pool,
-        tokenAmounts.map((amount) => amount.toString()),
-        '6300741387055771004078',
+        tokenAmounts,
+        '6094084206629046579738',
         true
       );
-      expect(priceImpact.toString()).to.eq('1584599872926409');
+      expect(priceImpact.toString()).to.eq('31775869758678519');
     });
   });
 });
