+++ balancer-js/src/modules/pools/pool-types/stablePhantom.module.ts
@@ -2,7 +2,7 @@ import { StablePhantomPoolExit } from './concerns/stablePhantom/exit.concern';
 import { StablePhantomPoolJoin } from './concerns/stablePhantom/join.concern';
 import { StablePoolLiquidity } from './concerns/stable/liquidity.concern';
 import { PhantomStablePoolSpotPrice } from './concerns/stablePhantom/spotPrice.concern';
-import { StablePhantomPriceImpact } from './concerns/stablePhantom/priceImpact.concern';
+import { StablePoolPriceImpact } from './concerns/stable/priceImpact.concern';
 import { PoolType } from './pool-type.interface';
 import {
   ExitConcern,
@@ -18,6 +18,6 @@ export class StablePhantom implements PoolType {
     public join: JoinConcern = new StablePhantomPoolJoin(),
     public liquidity: LiquidityConcern = new StablePoolLiquidity(),
     public spotPriceCalculator: SpotPriceConcern = new PhantomStablePoolSpotPrice(),
-    public priceImpactCalculator: PriceImpactConcern = new StablePhantomPriceImpact()
+    public priceImpactCalculator: PriceImpactConcern = new StablePoolPriceImpact()
   ) {}
 }
